+++ pkg/environment/utils.go
@@ -7,6 +7,9 @@ import (
 	"fmt"
 	"os"
 	"strconv"
+	"strings"
+
+	"github.com/onsi/ginkgo"
 )
 
 // Returns the integer value of an environment variable.
@@ -39,4 +42,33 @@ func MustGetEnv(key string) (string, error) {
 		return "", fmt.Errorf("missing required environment variable: %s", key)
 	}
 	return value, nil
+}
+
+func SetIfNotSet(key, value string, t ginkgo.GinkgoTInterface) {
+	if _, b := os.LookupEnv(key); !b {
+		if err := os.Setenv(key, value); err != nil {
+			t.Fatalf("Could not set environment variable %s", key)
+		}
+	}
+}
+
+func DefaultTestConfiguration(t ginkgo.GinkgoTInterface) {
+	SetIfNotSet(CatalogConnectorServiceAddressKey, "http://localhost:50085", t)
+	SetIfNotSet(VaultAddressKey, "http://127.0.0.1:8200/", t)
+	SetIfNotSet(EnableWebhooksKey, "false", t)
+	SetIfNotSet(ConnectionTimeoutKey, "120", t)
+	SetIfNotSet(MainPolicyManagerConnectorURLKey, "http://localhost:50090", t)
+	SetIfNotSet(MainPolicyManagerNameKey, "MOCK", t)
+	SetIfNotSet(LoggingVerbosityKey, "-1", t)
+	SetIfNotSet(PrettyLoggingKey, "true", t)
+}
+
+// GetSystemNamespace returns the namespace of control plane
+func GetSystemNamespace() string {
+	if data, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace"); err == nil {
+		if ns := strings.TrimSpace(string(data)); len(ns) > 0 {
+			return ns
+		}
+	}
+	return DefaultControllerNamespace
 }
